src/main/java/ec/gob/cj/pesnot/auditoria/ApiRestAuditoria4Application.java
package ec.gob.cj.pesnot.auditoria;

import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.boot.builder.SpringApplicationBuilder;
import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
import org.springframework.web.servlet.config.annotation.EnableWebMvc;

import springfox.documentation.swagger2.annotations.EnableSwagger2;


@EnableWebMvc
@EnableSwagger2
@SpringBootApplication
public class ApiRestAuditoria4Application  extends SpringBootServletInitializer {
	
	@Override
	protected SpringApplicationBuilder configure(SpringApplicationBuilder builder) {
	    return builder.sources(ApiRestAuditoria4Application.class);
	}

	public static void main(String[] args) {
		SpringApplication.run(ApiRestAuditoria4Application.class, args);
	}

}
